#3400 Remove unnecessary casts to long and double

+++ processor/src/main/java/org/mapstruct/ap/internal/util/NativeTypes.java
@@ -284,7 +284,7 @@ public class NativeTypes {
 
         @Override
         public Class<?> getLiteral() {
-            return float.class;
+            return double.class;
         }
 
     }
@@ -363,7 +363,7 @@ public class NativeTypes {
 
         @Override
         public Class<?> getLiteral() {
-            return int.class;
+            return long.class;
         }
     }
